app/requirements.txt
psycopg2-binary==2.9.9
fastapi==0.111.0
keydb==0.0.1
